Bugfix: do not load archive product template when shop page is not set
When the shop page is not set `get_option( 'woocommerce_shop_page_id' )` returns an empty string, which "tricks" `is_page()` to return true.  The WooCommerce helper function wc_get_page_id( 'shop' ) returns `-1` when the shop page is not set, and causes `is_page()` to correctly return false in that case.

Fixes #31